usb: phy: msm: Fix state machine worker logic for OTG mode.
Issue is that if the board is booted without micro usb cable and usb
device attached, kernel fails to populate the usb host and device.
The reason for this is that the state machine worker logic only checks
for USB_DR_MODE_PERIPHERAL and USB_DR_MODE_HOST modes to run worker
thread. However if the phy is configured in OTG mode it would fail
to run the state machine, resulting in failure to detect for very
first time.

This patch fixes by adding OTG checks in to the logic.

Issue is noticed on Qualcomm Dragon board DB410C.

Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@linaro.org>
(cherry picked from commit 15fddd0d508b672385f4fd6fed33c179a7943c5d)
Signed-off-by: Paolo Pisati <paolo.pisati@canonical.com>