UBUNTU: [packaging]: do not build host tools on hwe-edge
When syncing from the master kernel, do_tools_host = true for amd64, but we
don't want to build that. do_tools_common should be guarding from that, but for
now the rules are not doing that. So, we set it to false on hooks.mk, so we
won't try to build that package.

Ignore: yes

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>