UBUNTU: [Config]: enable CONFIG_ZRAM
BugLink: https://bugs.launchpad.net/bugs/1766823

zram is an useful module for memory constrained systems. However it's
imposibble for KVM kernel user to use it.

Enable the CONFIG_ZRAM and other related configs like in generic kernel to make
it available for KVM kernels, so that it can pass the ubuntu_zram_smoke_test
in the SRU regression test suite.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Kamal Mostafa <kamal@canonical.com>
Acked-by: Kleber Souza <kleber.souza@canonical.com>
Signed-off-by: Khalid Elmously <khalid.elmously@canonical.com>