UBUNTU: [Packaging] Update config after running update
The update script copies all config files from the
Bionic master and then has to remove the snapdragon
one as that is not built for hwe.
However that will always cause the next call to updateconfigs
to re-arrange some options which now can move to the common
config file.

Probably the update script needs to be changed to always do
an updateconfigs run after it has copied files.

Ignore: yes
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>